Added docstrings and inside method, changed between method.
The between method only worked on points in the first quadrant.
Fixed that. The inside method is similar to between with the
proviso that the target point has to be within all the lines
that can be constructed from the bounding points p and q. If
the probe point is on the line, it is not inside but it is
between.