Updated Point class with typing information.
Added type information to most of the Point class and provided some
typing abstractions into twod.types:

- Numeric:       float | int
- Coordinate:    Tuple[Numeric, Numeric]
- PointOrScalar
- PointType

I'm not sure how to apply typing to methods of classes that return
instances of that class or methods whose are arguments are instances
of this class. I punted and made a PointType which I think equates
to Any according to the docs I read. This will probably need to be
revisited.

Tests pass, have not added linting or type checking to pre-commit.